requirements.txt
Flask==2.0.1
firebase_admin==4.0.1
cryptography==3.1.1
python-dotenv==0.18.0
pyTelegramBotAPI==3.8.2